# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SmartResume AI Resume Analyzer API
  description: |
    ## SmartResume AI Resume Analyzer API

    Intelligent backend system that transforms raw resume documents into structured data and provides 
    personalized career guidance through AI-powered analysis.

    ### Features

    * **Document Processing**: Extract text from PDF, DOCX, and TXT files with OCR fallback
    * **Entity Extraction**: Use advanced NLP models to identify skills, experience, and qualifications
    * **Semantic Analysis**: Calculate compatibility scores between resumes and job descriptions
    * **AI Feedback**: Generate personalized career coaching recommendations using Google Gemini
    * **Secure Storage**: Store analysis results with user authentication and data protection

    ### Authentication

    All endpoints (except health checks) require JWT authentication via Supabase Auth.
    Include the JWT token in the Authorization header: `Bearer <token>`

    ### Rate Limits

    * Analysis endpoints: 10 requests per user per hour
    * Upload endpoints: 20 requests per user per hour
    * Other endpoints: 100 requests per user per hour

    ### File Upload Limits

    * Maximum file size: 10MB
    * Supported formats: PDF, DOCX, TXT
    * Text extraction confidence threshold: 0.80

    ### Response Times

    * 95% of requests complete within 30 seconds
    * Average analysis time: 8-15 seconds
    * File upload processing: 2-5 seconds
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /upload:
    post:
      tags:
        - Upload
      summary: Upload resume document
      description: |
        Upload and process a resume document in PDF, DOCX, or TXT format.
        The system will extract text content and store the processed resume.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Resume file (PDF, DOCX, or TXT, max 10MB)
              required:
                - file
      responses:
        '200':
          description: File uploaded and processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Invalid file format or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unsupported_format:
                  summary: Unsupported file format
                  value:
                    error_code: "UNSUPPORTED_FORMAT"
                    message: "File format 'image/jpeg' is not supported"
                    details:
                      supported_formats: ["application/pdf", "application/vnd.openxmlformats-officedocument.wordprocessingml.document", "text/plain"]
                      detected_format: "image/jpeg"
                    timestamp: "2023-11-04T10:30:00Z"
                    request_id: "uuid-123"
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Document processing failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /resumes:
    get:
      tags:
        - Upload
      summary: Get user's uploaded resumes
      description: Returns a list of all resumes uploaded by the authenticated user
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: List of user's resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResumeListItem'
...
  /analyses:
    get:
      tags:
        - History
      summary: Get analysis history
      description: Retrieve paginated list of user's past analyses
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: sort_by
          in: query
          description: Sort field
          schema:
            type: string
            enum: [created_at, match_score]
            default: created_at
        - name: sort_order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Analysis history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisListResponse'
  /analyses/{analysis_id}:
    get:
      tags:
        - History
      summary: Get specific analysis
      description: Retrieve detailed results for a specific analysis
      parameters:
        - name: analysis_id
          in: path
          required: true
          description: Analysis identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Analysis details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisDetailResponse'
        '404':
          description: Analysis not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
  schemas:
...
    UploadResponse:
      type: object
      properties:
        resume_id:
          type: string
          format: uuid
          description: Unique identifier for the uploaded resume
        file_name:
          type: string
          description: Original filename
        file_size:
          type: integer
          description: File size in bytes
        processing_method:
          type: string
          enum: [pdfplumber, ocr, docx, text]
          description: Method used to extract text
        confidence_score:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence score of text extraction
        text_length:
          type: integer
          description: Length of extracted text
        uploaded_at:
          type: string
          format: date-time
          description: Upload timestamp
      required:
        - resume_id
        - file_name
        - file_size
        - processing_method
        - confidence_score
        - text_length
        - uploaded_at
      example:
        resume_id: "123e4567-e89b-12d3-a456-426614174000"
        file_name: "john_doe_resume.pdf"
        file_size: 245760
        processing_method: "pdfplumber"
        confidence_score: 0.95
        text_length: 2847
        uploaded_at: "2023-11-04T10:30:00Z"
    AnalysisResponse:
      type: object
      properties:
        analysis_id:
          type: string
          format: uuid
          description: Unique identifier for the analysis
        match_score:
          type: number
          minimum: 0
          maximum: 100
          description: Compatibility score (0-100)
        ai_feedback:
          $ref: '#/components/schemas/AIFeedback'
        matched_keywords:
          type: array
          items:
            type: string
          description: Keywords found in both resume and job description
        missing_keywords:
          type: array
          items:
            type: string
          description: Important keywords missing from resume
        processing_time:
          type: number
          description: Analysis processing time in seconds
        created_at:
          type: string
          format: date-time
          description: Analysis timestamp
      required:
        - analysis_id
        - match_score
        - ai_feedback
        - matched_keywords
        - missing_keywords
        - processing_time
        - created_at
    AIFeedback:
      type: object
      properties:
        overall_assessment:
          type: string
          description: Overall assessment of the resume-job match
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
        strengths:
          type: array
          items:
            type: string
          description: Identified strengths in the resume
        priority_improvements:
          type: array
          items:
            type: string
          description: High-priority areas for improvement
      required:
        - overall_assessment
        - recommendations
        - strengths
        - priority_improvements
    Recommendation:
      type: object
      properties:
        category:
          type: string
          enum: [skills, experience, education, formatting, keywords]
          description: Category of the recommendation
        priority:
          type: string
          enum: [high, medium, low]
          description: Priority level of the recommendation
        suggestion:
          type: string
          description: Specific improvement suggestion
        impact:
          type: string
          description: Expected impact of implementing the suggestion
      required:
        - category
        - priority
        - suggestion
        - impact
    ResumeListItem:
      type: object
      properties:
        resume_id:
          type: string
          format: uuid
        file_name:
          type: string
        text_length:
          type: integer
        uploaded_at:
          type: string
          format: date-time
      required:
        - resume_id
        - file_name
        - text_length
        - uploaded_at
    AnalysisListResponse:
      type: object
      properties:
        analyses:
          type: array
          items:
            $ref: '#/components/schemas/AnalysisListItem'
        total_count:
          type: integer
          description: Total number of analyses for the user
        page:
          type: integer
          description: Current page number
        page_size:
          type: integer
          description: Number of items per page
        has_next:
          type: boolean
          description: Whether there are more pages
      required:
        - analyses
        - total_count
        - page
        - page_size
        - has_next
    AnalysisListItem:
      type: object
      properties:
        analysis_id:
          type: string
          format: uuid
        job_title:
          type: string
        match_score:
          type: number
          minimum: 0
          maximum: 100
        created_at:
          type: string
          format: date-time
        processing_time:
          type: number
      required:
        - analysis_id
        - job_title
        - match_score
        - created_at
        - processing_time
    AnalysisDetailResponse:
      allOf:
        - $ref: '#/components/schemas/AnalysisResponse'
        - type: object
          properties:
            job_description:
              type: string
              description: Original job description used for analysis
...
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error context
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          description: Unique request identifier for tracing
      required:
        - error_code
        - message
        - timestamp
        - request_id
...